health_api/requirements.txt
fastapi
uvicorn
python-dotenv
supabase
psycopg2-binary